Split out SystemTestStateStoreFakeCommitsThroughputTest
